+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.5]
+
+- use `selectedValue` instead of `value` for `khao-ui-selectfield`
+
+## [3.1.4]
+
+- use `min-height` instead of `height` for `form-field`
+
 ## [3.1.3]
 
 - use fatter icons for `navigate-back` and `navigate-forward`